moved file-io into a separate directory
