+++ src/analytical_solutions/sedov_solution/main.cpp
@@ -33,10 +33,9 @@
 #include <iomanip>
 #include <cmath>
 #include <vector>
-#include <filesystem>
 
-#include "arg_parser.hpp"
-#include "file_utils.hpp"
+#include "io/arg_parser.hpp"
+#include "io/file_utils.hpp"
 
 #include "sedov/sedov_generator.hpp"
