+++ src/io/ifile_writer.hpp
@@ -3,6 +3,7 @@
 #include <string>
 #include <vector>
 
+#include "file_utils.hpp"
 #include "mpi_file_utils.hpp"
 #include "particles_data.hpp"
